Update application to a Pokedex with infinite scrolling
The application's layout and title were updated to a Pokedex. Added new dependencies such as framer-motion and use-debounce. The home page now allows fetching and displaying Pokemon on an infinite scroll. New components such as Search, Navbar, LoadPokemon, and PokemonCard were added to support this functionality. A new interface for Pokemon was also added.